requirements.txt
flask
numpy
opencv-python
tensorflow
pandas
watchdog
pillow
scikit-learn
keras

